+++ src/pylinkage/optimization/__init__.py
@@ -2,9 +2,13 @@
 
 __all__ = [
     "collections",
+    "differential_evolution_optimization",
+    "differential_evolution_optimization_async",
     "generate_bounds",
     "kinematic_maximization",
     "kinematic_minimization",
+    "minimize_linkage",
+    "minimize_linkage_async",
     "OptimizationProgress",
     "particle_swarm_optimization",
     "particle_swarm_optimization_async",
@@ -17,6 +21,12 @@ from .async_optimization import (
     OptimizationProgress as OptimizationProgress,
 )
 from .async_optimization import (
+    differential_evolution_optimization_async as differential_evolution_optimization_async,
+)
+from .async_optimization import (
+    minimize_linkage_async as minimize_linkage_async,
+)
+from .async_optimization import (
     particle_swarm_optimization_async as particle_swarm_optimization_async,
 )
 from .async_optimization import (
@@ -27,6 +37,12 @@ from .grid_search import (
 )
 from .particle_swarm import (
     particle_swarm_optimization as particle_swarm_optimization,
+)
+from .scipy_optimize import (
+    differential_evolution_optimization as differential_evolution_optimization,
+)
+from .scipy_optimize import (
+    minimize_linkage as minimize_linkage,
 )
 from .utils import (
     generate_bounds as generate_bounds,